+++ src/index/iterator/mod.rs
@@ -1,7 +1,10 @@
 mod children;
-mod gosper;
 mod grid_path;
 
 pub use children::Children;
-pub use gosper::Gosper;
 pub use grid_path::GridPathCells;
+
+#[cfg(feature = "geo")]
+mod gosper;
+#[cfg(feature = "geo")]
+pub use gosper::Gosper;