+++ src/index/mod.rs
@@ -9,7 +9,6 @@ mod vertex;
 
 pub use cell::CellIndex;
 pub use edge::{DirectedEdgeIndex, Edge};
+pub use iterator::{Children, GridPathCells};
 pub use mode::IndexMode;
 pub use vertex::{Vertex, VertexIndex};
-
-pub use iterator::{Children, Gosper, GridPathCells};